pyproject.toml
[project]
name = "o-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "pydantic-settings>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "alembic>=1.12.0",
    "fastapi-mail>=1.5.0",
    "python-dotenv>=1.1.1",
]
